Embeddings: bailout early for zero results (#53024)
This adds early bailouts when we know we can't return any results. 

This fixes an issue where searching zero repositories would actually
search all repositories. The problem was that `repo:^$` actually returns
all repositories. This is a search bug that is unrelated to this
endpoint and is not fixed by this PR.