cody-gateway/httpapi: refactor upstreamHandler methods into parameter struct (#52967)
Readability improvement - there are a lot of callbacks and a lot of
parameters to `makeUpstreamHandler`, when reading it from the callsite
e.g. `newAnthropicHandler` it's a bit difficult to figure out which is
which without knowing the function signatures well. Turning these
parameters into a single options struct parameter gives them names at
the callsite.

## Test plan

CI