perforce: Perform changelist mapping for new and existing depots imported as git repos (#51860)
Part of #40330.

## Why

For a perforce depot, we want to be able to answer the question: "What
is the changelist ID when this change was made?". In a previous PR
(#51872) we added the DB layer to support this
(`repo_commits_changelists` table). And in this PR we add the job
management pipeline and the job execution itself.

## What

A new "plugin" / "service" named `Perforce` is added to
gitserver.Server. It is responsible for all things perforce and exists
as a sub-package of `cmd/gitserver/server`. This makes it easy to break
out of gitserver if needed in the future and also allows to be tested
independently of gitserver.

For every perforce depot, the job management pipeline will:

1. For a new repo, map all the commits to their corresponding changelist
IDs in the DB
2. For existing repos, when a repo is synced all commits will be mapped
to their corresponding changelist IDs in the DB
3. For new commits (changelists) synced to a repo, the newly discovered
commits will be mapped to the DB

## Caveats

1. If a repo is deleted from the DB, we will drop all rows from the
`repo_commits_changelists` table at the moment
2. If a repo is "deleted" from the code host, we do not touch the
entries in the `repo_commits_changelists` table. This means if a repo is
corrupted and users want to sync a repo from a clean slate, we won't be
overwriting the existing rows. This needs to be handled gracefully and
will follow up in a subsequent PR
3. If a repo has a force push, the last mapped commit in the DB may no
longer be a valid revision. This will mean the `lastMappedCommit..HEAD`
range will no longer be valid. This needs to be handled gracefully and
will follow up in a subsequent PR

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>