Search: enable cross-repo ranking by default (#53031)
This PR updates the `search-ranking` feature flag to default to 'true'.
This greatly improves the result ranking search ranking for cross-repo
searches, or for very large repos.

We feel ready to GA this since it's been running successfully on dot com
and S2 for several months, and had a successful customer trial.

Note: this enables only **part** of the "Intelligent ranking" feature.
The code intel jobs to produce file ranks are still not enabled by
default. We'll make a separate "GA" decision for those jobs.