Cody-gateway - Dotcom user actor source  (#52813)
This adds a new actor source for the cody gateway that accepts verified
dotcom users and allows them to utilize chat, completions and
embeddings.

The main use case for this is so that App will be able to talk directly
to the cody-gateway to avoid App users needing to provide their own api
keys and knowledge of completions/embeddings apis.

Access:

This actor uses a derived key generated from a user api token, this
allows disable gateway access by deleting the api token. I did not
expose an endpoint to create these cody-gateway specific user tokens
because the main use case is for the Cody app, which runs the
sourcegraph backend and can generate this token directly from the api
token the user created to connect App to dotcom.

Rate limits:

- **Chat & Code completions** utilize the rate limits that have defaults
on dotcom and also allows an override per user. This rate limit is
separate from the limit imposed when on dotcom because it is imposed
directly at the gateway. Effectively this means any user can get 2x the
limit split evenly between dotcom usage and Cody App usage.
- **Embeddings** utilize a hard coded default "lifetime" limit as was
previously determined