RFC settings: introduce Service interface
This is a stub interface which is a proposal for how we use the settings
API. The intention is to remove the currently exported functions. From
reading call sites on the existing functions this should be easy to
introduce.

Please look at the code which has comments and please give feedback.
Based on that I will follow up with implementing it or closing.

Test Plan: n/a