Cody: Track the range non-stop fixup is editing (#53147)
In classic Fixup, you could not touch the selection while Cody was
working. The premise of non-stop Cody is that you never have to stop
editing while Cody is doing fixups. So as you edit a file, we need to
keep track of the range in the file Cody's diffs should stream into.
This does that tracking.

## Test plan

````
pnpm test:unit -- tracked-range.test.ts
pnpm test:unit && pnpm test:integration && pnpm test:e2e
````

Manual test:

1. Set up Cody; set `"cody.experimental.nonStop": true` in user JSON.
2. Open a file with a bunch of code
3. Select a function in the middle of the file and Cody: Fixup
(Experimental), instruct Cody to make some change

Then, while the edits are streaming in, and after they are done:

- Add or remove lines before where Cody is editing, then edit within the
region Cody is editing
- Edit in the region where Cody is editing with and without causing
conflicts
- Add or remove lines after where Cody is editing, then edit within the
region Cody is editing

The CodeLens should move around so it appears in the right place in the
file.

Edits before/after the region Cody is editing don't trigger diff
recalculation; we are simply relying on vscode to shift our decorations
around. So to trigger diff recalculation to see if the range tracking
was effective, briefly edit within the region Cody edited to trigger
diff recalculation.

Before this change we don't track the range so the test described above
reports many spurious conflicts because the diff is out of alignment.
After this change, the recomputed diff is accurate.