executors: Fix Heartbeat Backwards Compatibility (#53176)
When Executors is 5.1 (newer) and Sourcegraph is older (5.0), the
executor stopped sending the `version` in the V1 request. This caused
the older Sourcegraph to send a response payload that is a 4.x format
(e.g. `[123]`). This caused unmarshalling errors. We need to add the
version `V2` to the request payload to ensure Sourcegraph sends a
structured JSON instead of an array.

## Test plan

Updated existing tests.