Move concurrency limit into RateLimit type (#53182)
This makes it so that we can do special computations based on the
source, and also potentially send overwrites for concurrency from the
dotcom side later. This is mostly in preparation for separating
interactive and background embedding jobs, to make sure that interactive
requests stay functional and allow higher concurrency, while we will be
way stricter about the heavy bulk jobs.

## Test plan

Tests are still passing.