Embeddings: remove context detection index (#53131)
Our context detection embeddings index adds complexity to the system:
* We need to make sure it's populated once when embeddings are
configured. **Currently this is buggy ... for example it's totally
missing if an admin sets an embeddings policy (but never manually
schedules)!**
* When the embeddings model changes, admins will need to migrate it to
the new model too. We'd need to add some automatic upgrade, or surface
this index in the admin UI.

The complexity probably outweighs the benefit we're getting from it.
LLMs do a good job of ignoring irrelevant context, so supplying context
when it's not needed won't hurt response quality.

The main risk: chat fluency might get worse. We might be worse at
detecting when a chat message is referencing the previous one, so it
should use the existing context. I tried to mitigate this by expanding
the "no context regexes" with a bunch of common cases I've encountered.