Completions: Fix issue where completion sometimes starts in a new line (#53246)
I noticed a small issue when completions are triggered for non-empty
lines.

We have one of the three requests that injects a `\n` at the end of the
prefix. This was added back _after_ the trimming logic which of course
means that it would include the `\n` in the result, not something that
we want.

We have to evaluate and see if this logic really makes sense. For now
it's just a way to "trigger" a response.

## Test plan

- Tested this locally

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->