[license-checks] add check for anomalous license usage (#53373)
## Description

Send a slack message to a channel if we figure out, that the 48-hour p50
time difference between license-check calls for the same site_id and
license key pair is less than 10 hours.

Achieved by adding events to event_logs and then on a schedule checking
the time diff with an sql query.

This check will only execute on dotcom if the
`slackLicenseAnomallyWebhook` site config option is defined. If there is
no webhook config, nothing will be checked, as there is no receiver of
the result.

## Test plan

unit tests + tested locally

This will only be enabled on dotcom, so blast radius is low.