Embeddings: add 'model version' to indexes (#53264)
This PR updates the embeddings indexes to include the model used to
generate the indexes. With this information, we now
* Throw a search error if the current embeddings config does not match
the model used to generate the index. This avoids bad situations like a
dimension mismatch between query and index vectors, or silently
returning poor results.
* When reembedding a repo, always fall back to a full index if the
embeddings model has changed.

To support this, the PR adds the concept of "index format version" to
embeddings indexes. This lets us start to encode the embeddings model,
while still handling old indexes. This is generally useful to have, as
it lets us add other metadata or change the index format in the future.

Addresses #52668