Return files modified by specified contributors (#53206)
A [user
requested](https://sourcegraph.slack.com/archives/C02UM8A9M2R/p1678812044400949)
being able to use `author:` keyword along with a pattern to search for
file matches. Currently `author` keyword is valid for commit and diff
search types. The use case is that a provided pattern will match a part
of the file that was modified by the author and that matching line **may
or may not** be part of the modified lines of the author's commit.

To support this we add a new predicate `file:has.contributor()`. Some
characteristics of the predicate:

- _Contributor_ is an existing
[struct](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/gitserver/gitdomain/common.go?L208).
- Regex is supported. We use same regex expression
[logic](https://sourcegraph.com/github.com/sourcegraph/sourcegraph@2a63d0dc6ff8f520ff26a91c65b4a1366cf36712/-/blob/internal/gitserver/search/match_tree.go?L23-25)
as `author`.
- Case sensitivity can be applied for the contributor criteria of
`file:has.contributor()` with the existing url parameter.
- Negation is supported.
- Multiple predicates will be AND'ed together by default if boolean
operator not specified. This is how `has.owner` works, not sure if all
of the predicates are aligned on this. General discrepancies are tracked
and described in detail
[here](https://github.com/sourcegraph/sourcegraph/issues/50539).
- Filtering results happens **after** pattern matches are streamed from
child job
([discussion](https://github.com/sourcegraph/sourcegraph/issues/50880#issuecomment-1561650483).
For this reason a user might not always find what they're looking for if
the pattern matches only an older revision of a file and that older
revision is not the revision streamed back from the child job. [One of
my
proposals](https://github.com/sourcegraph/sourcegraph/issues/50880#issuecomment-1560170183)
would resolve this by iterating through batches of commits with
`commitSearcher` but I'm not sure about performance impact of this
approach.

[GH Issue](https://github.com/sourcegraph/sourcegraph/issues/50880) 

## Test plan

new tests for `FileHasContributorsJob`