mvu: handle patch-only upgrade for autoupgrade (#53489)
If attempting to do an upgrade between patch versions with autoupgrade,
we would fail with a schema out of date error. This happens because when
checking whether we need to perform an upgrade, we only check
major/minor versions.
This PR does an additional check whether a patch version upgrade needs
to be performed, if we don't need to perform a larger upgrade. We
attempt to perform the larger upgrade flow for patch versions, as
creating the upgrade range will error when major+minor versions are
equal

## Test plan

Did local upgrade from 5.0.0 to 5.0.5