Embeddings: remove the query embeddings cache (#53506)
Now that we've removed the context detection index, caching query
embeddings is no longer critical. Before, it was helpful because we'd
embed the same query twice, once for context detection and once for
retrieving similar files. Also, the cache does not key on the embeddings
model, so it could return incorrect results when the model changes.

Follow up to #53131