cody: use embeddings for completions (#53223)
- POC embeddings for code completions.
- Adds two code matches and `.md` match to the code completions prompt.
- Embeddings matches take precedence over matches found via Jaccard
distance (we discard local matches if an embeddings match exists for the
same filename)
- Does not add additional latency to code completions. We fetch
embeddings in the background for a file that the user started editing
and cache results for 5 mins. If embeddings are available on the
subsequent completions call, they are used along with local context
matches.
- I do not filter embedding matches based on the current language yet.
Based on local testing, the embeddings search query always returns
relevant results. We can add this logic in a follow-up after
more dogfooding.
- Building on top of
https://github.com/sourcegraph/sourcegraph/pull/53205