App: Rename to Cody in app build scripts (#53552)
Rename to Cody in the build scripts that build the app. This should only
affect the file names of the releases.

My editor automatically applied formatting to the shell files, so that's
where the minor whitespace changes come from.

This PR only touches the build scripts. The other parts of the rebrand
are in:

- #53518.

This is part of:

- #53120 

## Test plan

- Build app
- Verify that it works and that the releases are named "Cody" where they
were previously named "Sourcegraph"