App - Use git remote as repo "name" if available (#53610)
The editor extensions attempt to use the git remote to determine if the
repo is available indexed, servegit was using the folder name as the
repo name. This caused extension users to need to manually configure the
cody codebase settings in vs code. This change updates servegit to use
the git remote as the name if there is one available and the fall back
to the folder name if not.