Embeddings: check embeddings response length (#53716)
We've had a few situations where we generate corrupt indexes because we
trust that the embeddings service is returning the correct number of
embeddings. In the past, this has failed when the model changes
mid-embed (and changes the number of dimensions) and when the embeddings
generation service/proxy is misbehaving and returning an empty response
with a successful status code.

This moves the error to index time and makes it nice and clear rather
than depending on the cryptic panics we get during search time.

(cherry picked from commit e4996bd9dfd8887ebf0265de57a7d8d1dfa4ef81)