gitserver: Add recordingfactory to Fetch of git vcs syncer
In this commit we wrap the fetch command in a RecordingCommandFactory
similar to what is currently done at the call site of CloneCommand.

In order to make this cleaner - we add the RecordingCommandFactory on
the GitRepoSyncer and introduce a constructor - NewGitRepoSyncer to
ensure that the factory is always initialised and make GitRepoSyncer
private.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>