search: temporarily record stats values
I noticed in some places we use "len(repos.Minimal)" to indicate the
number of indexed repos. However, in practice this should be the same as
Stats.Repos. The only time Stats.Repos should be inaccurate is when
rebalancing between zoekt shards.

I am going to collect data on sourcegraph.com via these metrics to see
what happens in practice. I am recording two ways to calculate as well
as if we have any crashes (which may also not be reported correctly in
some places)

Test Plan: CI