wrexec: handle unsetting GitRecorder config
Minor issue I came across while reading code. Previously if a site-admin
removed GitRecorder configuration that would be ignored until the next
process restart. This explicitly handles that.

Additionally a test is added that we handle nil predicates. We already
relied on this behaviour since if GitRecorder was not configured we
would never call update => we had a nil predicate.

Test Plan: CI