guardrails: do case sensitive search (#53771)
We want to do this since we are currently looking for exact matches.
Additionally attribution searches are slow on dotcom. My working theory
for why is related to how we lookup ngrams. By specifying case
sensitivity we drastically reduce the number of ngrams we need to check.

We also switch the protocol from Batch to Streaming. In practice this
makes no difference since we specify count. Using streaming is more
consistent with the other search clients.

Test Plan: go test