App: Fix paper cuts in UI and rename sourcegraph to cody app (#53714)
Closes https://github.com/sourcegraph/sourcegraph/issues/53629
Closes https://github.com/sourcegraph/sourcegraph/issues/53694
Closes https://github.com/sourcegraph/sourcegraph/issues/53699
Closes https://github.com/sourcegraph/sourcegraph/issues/52416
Closes https://github.com/sourcegraph/sourcegraph/issues/53700
Closes https://github.com/sourcegraph/sourcegraph/issues/53708

## TODO 
- [x] The App system tray icon should be updated to the Cody Icon. 
- [x] I still see the Sourcegraph logo in the top lefthand corner. We
probably don't need this anymore, since we have the "Cody AI" tab and
they direct to the same chat view.
- [x] In the settings dropdown in the app UI there's a "Teams" option
that doesn't go anywhere. I don't know what this is or where it's
supposed to go.
- [x] The cloud icon in the app UI still always says "indexing." This is
pretty prominent UI and (I believe) refers to Code Search indexing, not
embeddings jobs. If it isn't doing anything, we should remove it.
- [x] Update the Authorization page to say "Cody App" instead of
"Sourcegraph App" when connecting App to dotcom.

## Test plan
- Check the points from the todo section