app: clicking on dock icon should open main window (#53793)
Closes https://github.com/sourcegraph/sourcegraph/issues/53756

This allows clicking the dock icon to show the main window if it is
closed but the app is still running.

This is a temporary solution that only works if the app has a single
window. If we need to add more windows in the future, we need to wait
until https://github.com/tauri-apps/tauri/issues/3084 is fixed.

## Test plan

1. Open app
2. Close with red button on top-left of window
3. Click on Cody icon in dock

App main window should be shown again.