cody: add maxPromptToken settings to site-config
This adds 3 new values to the site-config under `completions`:

- `fastChatModelMaxPromptTokens`
- `chatModelMaxPromptTokens`
- `completionModelMaxPromptTokens`

These values don't have an effect on the server-side of things (they're
not used to truncate messages to LLM for example). They're hints to the
clients as to how many tokens they should use max when talking to a
configured LLM.

The change in here also exposes the values in GraphQL, along with the
model names. They can be retrieved like this:

```graphql
query {
  site {
    codyLLMConfiguration {
      chatModel
      chatModelMaxPromptTokens
      fastChatModel
      fastChatModelMaxPromptTokens
      completionModel
      completionModelMaxPromptTokens
    }
  }
}
```

If the values aren't set in the site-config, they are `null` and the
client needs to determine limit on its own.

Note:
- Specifically called with `LLMConfiguration` because I think
  `completions` is confusing and we should start moving away from it.
- Strictly speaking we don't need to expose the model names, but I do
  think it's better for error messages in the client.