search: introduce ExecutionResult to enforce telemetry collection
This just introduces the type without actually recording the telemetry.
Additionally we introduce a new parameter to Plan to capture the
subsystem. The goal is to start recording prometheus, honeycomb and
logging across all search clients. Currently we only record this for
GraphQL and streaming http requests.

This commit doesn't actually implement the telemetry sending. That will
come in the next (much smaller) commit.

Additionally code monitors does not yet support telemetry since it
directly uses the job interface. That will be followed up with.

Test Plan: go test