autoupgrade: use heartbeats to detect crashed autoupgrade attempts (#53859)
If a frontend crashes due to OOM or any other reason, in a way that it
fails to set the `success` column in `upgrade_logs` to `false`, then all
frontends will block trying to claim the autoupgrade lock until the
status is manually set in postgres (or the table is truncated). With a
heartbeat mechanism, we can allow claiming the lock if the last lock
hasnt received a heartbeat in a certain amount of time (here we're
giving 6 missed heartbeats space).

## Test plan

Added unit tests for the claim query, tested in docker-compose

![image](https://github.com/sourcegraph/sourcegraph/assets/18282288/5e79817a-10d9-452f-bcab-e125fc8a8333)

(cherry picked from commit 746a323e26615f23c14564b58460fa30868d68ac)