cody site-admin: fix refetch for one-time jobs
When a user clicks on "Schedule Embedding", we schedule the repos for
embedding and refetch the job list. However, the current refetch logic
fetches only n results, where n is the number of jobs currently
displayed. This leads to a confusing behavior which
the user can fix with a page refresh. Not ideal.

With this PR we try to fetch to a full page (10) of jobs which makes
more sense, because the number of jobs likely will have increased after
the user clicked on "Schedule Embedding".