Tracing: inherit from raw OTel trace for TraceFromContext (#53874)
Currently, if we have a raw OTel trace in the context but no active
`trace.Trace`, we `TraceFromContext` will not use that, so we can lose
some tracing information. This provides a temporary fix to make
`TraceFromContext` create a new, ephemeral `*Trace` that contains the
OTel span if it exists.

I think the real solution here is to just go all in on the OTel tooling
and make our custom trace stuff conform to the OTel interfaces. I'll
work on that in the background, but I want to get this out before 5.1.

(cherry picked from commit 8f2903f1bacc6d0ed497bbafbdb3742e57b1b7ce)