cody site-admin: fix refetch for one-time jobs (#53873)
When a user clicks on "Schedule Embedding", we schedule the repos for
embedding and refetch the job list to update the view. However, the
current refetch logic fetches only `n` results, where `n` is the number
of jobs currently displayed. This leads to a confusing behavior (1 job
visible, user schedules 3 more => 1 job visible), which the user can fix
with a page refresh. Not ideal.

With this PR we always fetch a full page, i.e. 10 jobs, which makes more
sense, because the number of jobs likely will have increased after the
user clicked on "Schedule Embedding".

The root cause of the bug is that `refetchAll` closes over the current
list of jobs, which I believe is the intention for the other call sites.
I don't think it is a good idea to modify `refetchAll` during the code
freeze. Returning a new function `fetchFirst` keeps the surface of this
change small.

## Test plan
CI

(cherry picked from commit 813571fe61bb6e7306b0b50b9fc578ae618903ff)