App - Ensure embeddings eventually get created for all repos added during setup (#53860)
During testing it was seen that occasionally enough time was not
allocated to allow repos to clone and therefore they did not have an
embeddings jobs scheduled.

This extends the amount of time the mutation is willing to wait to
attempt to schedule the initial embeddings job for newly added repos.

It also creates a default policy to keep embeddings up to date for all
repos. This should ensure that it is eventually consistent even if an
initial embedding job can't be scheduled.