App - Overall embeddings progress during app startup (#53909)
This adds a special embeddings progress that can be used by app during
the setup wizard. This differs in that it is tracking overall progress
expecting that every repo added should have embeddings generated. It
does not necessarily care about in progress or queued jobs if a previous
job as already completed.

Co-authored-by: William Bezuidenhout <william.bezuidenhout@sourcegraph.com>