cody: add maxPromptToken settings to site-config (#53850)
This adds 3 new values to the site-config under `completions`:

- `fastChatModelMaxTokens`
- `chatModelMaxTokens`
- `completionModelMaxTokens`

These values don't have an effect on the server-side of things (they're
not used to truncate messages to LLM for example). They're hints to the
clients as to how many tokens they should use max when talking to a
configured LLM.

The change in here also exposes the values in GraphQL, along with the
model names. They can be retrieved like this:

```graphql
query {
  site {
    codyLLMConfiguration {
      chatModel
      chatModelMaxTokens
      fastChatModel
      fastChatModelMaxTokens
      completionModel
      completionModelMaxTokens
    }
  }
}
```
```json
{
  "data": {
    "site": {
      "codyLLMConfiguration": {
        "chatModel": "anthropic/claude-v1",
        "chatModelMaxTokens": 10000,
        "fastChatModel": "anthropic/claude-instant-v1",
        "fastChatModelMaxTokens": null,
        "completionModel": "anthropic/claude-instant-v1",
        "completionModelMaxTokens": 5000
      }
    }
  }
}
```

If the values aren't set in the site-config, they are `null` and the
client needs to determine limit on its own.

Note:
- Specifically called with `LLMConfiguration` because I think
`completions` is confusing and we should start moving away from it.
- Strictly speaking we don't need to expose the model names, but I do
think it's better for error messages in the client.

## Test plan

- Manual testing