docker: fix prom-gcp being built with wrong base version (#53946)
I'm not sure to understand why our previous script ended up using the
wrong base image, but this PR fixes it by hardcoding it instead of
relyin on a `FROM ${BASE_IMAGE}` that ended up with the default instead.

As a result, prom-gcp was using 2.45 instead of 2.32, which doesn't have
the `--export.match` flag that was discussed in
[Slack](https://sourcegraph.slack.com/archives/C04MYFW01NV/p1687410706995679)

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Build it locally, double checked prom-gcp uses 2.35 and 2.42. 
Will push a main-dry-run to check the final image once normal CI is
green.