gitserver: Add recordingfactory to Fetch of git vcs syncer (#53741)
In this commit we wrap the fetch command in a RecordingCommandFactory
similar to what is currently done at the call site of CloneCommand.

In order to make this cleaner - we add the RecordingCommandFactory on
the GitRepoSyncer and introduce a constructor - NewGitRepoSyncer to
ensure that the factory is always initialised and make GitRepoSyncer
private.

## Note

The recording factory is initialised with a nil `shouldRecord` at the
moment so this won't start recording fetch commands immediately. We
update the factory similar to that in the server's handler and even then
it should only affect those repos that are listed in the conf. So in
addition to recording the clone path, now we will also record all
fetches of those repos - which could be a source of more writes to redis
(the backing store of the recording factory).

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>